+++ nautilus_trader/system/kernel.py
@@ -340,9 +340,7 @@ class NautilusKernel:
             msgbus=self._msgbus,
             cache=self._cache,
             clock=self._clock,
-            portfolio_bar_updates=(
-                (config.exec_engine and config.exec_engine.portfolio_bar_updates) or True
-            ),
+            config=config.portfolio,
         )
 
         ########################################################################